pyproject.toml
[tool.poetry]
name = "agentic-stock-analyzer"
version = "0.1.0"
description = "A Python-based stock analysis tool leveraging LangGraph, Gemini, and yfinance"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}, {include = "api"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pydantic = ">=2.0.0"
typing-extensions = ">=4.10.0"
fastapi = ">=0.100.0"
uvicorn = ">=0.34.0"
sse-starlette = ">=1.6.1"
python-dotenv = ">=1.0.0"
langchain-core = ">=0.3.83"
langchain-community = ">=0.4.0"
langchain-google-genai = ">=2.0.0"
langchain-anthropic = ">=0.3.0"
langchain-groq = ">=0.3.0"
langchain-ollama = ">=0.3.10"
langgraph = ">=0.3.0"
langgraph-checkpoint-sqlite = ">=2.0.0"
langsmith = ">=0.4.37"
ollama = ">=0.6.1"
yfinance = ">=0.2.40"
pandas = ">=2.0.0"
numpy = ">=1.26.2"
openpyxl = ">=3.1.2"
pyxlsb = ">=1.0.10"
httpx = ">=0.28.0"
requests = ">=2.32.0"
tenacity = ">=9.0.0"
click = ">=8.1.0"
attrs = ">=24.0.0"
PyYAML = ">=6.0.1"
python-dateutil = ">=2.8.2"
duckduckgo-search = ">=6.1.7"
edgartools = ">=3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
cache_dir = "/tmp/.pytest_cache"
python_files = ["test_*.py"]

[tool.ruff]
target-version = "py310"
line-length = 120
